requirements.txt
google-generativeai==0.4.0
pillow==10.2.0
streamlit
streamlit-option-menu==0.3.12
altair==4.2.2